Update: finish scaleup
